Heavy metal biomonitoring in Galicia

This data-set relates to two studies on lead concentration in moss samples, in micrograms per gram dry weight, collected in Galicia, norther Spain. The data are from two surveys, one conducted in July 2000. The variables are as follows:

x x-coordinate of the spatial locations.

y y-coordinate of the spatial locations.

lead number of tested people for the presence nodules.

The coordinate reference system of the data is 32629.

Usage

data(galicia)

Format

A data frame with 195 rows and 4 variables

Source

Diggle, P.J., Menezes, R. and Su, T.-L. (2010). Geostatistical analysis under preferential sampling (with Discussion). Applied Statistics, 59, 191-232.