River-blindness in Liberia

This data-set contains counts of reported onchocerciasis (or riverblindess) cases from 91 villages sampled across Liberia. The variables are as follows:

lat latitude of the of sampled villages.

long longitude of the sampled villages.

ntest number of tested people for the presence nodules.

npos number of people that tested positive for the presence of nodules.

elevation the elevation in meters of the sampled village.

log_elevation the log-transformed elevation in meters of the sampled village.

Usage

data(liberia)

Format

A data frame with 90 rows and 6 variables

Source

Zouré, H. G. M., Noma, M., Tekle, Afework, H., Amazigo, U. V., Diggle, P. J., Giorgi, E., and Remme, J. H. F. (2014). The Geographic Distribution of Onchocerciasis in the 20 Participating Countries of the African Programme for Onchocerciasis Control: (2) Pre-Control Endemicity Levels and Estimated Number Infected. Parasites & Vectors, 7, 326.