Plot AnPIT Results from Model Validation

This function plots AnPIT results from a `RiskMap.spatial.cv` object obtained using the `assess_pp` function.

Usage

plot_AnPIT(
  object,
  mode = "average",
  test_set = NULL,
  model_name = NULL,
  combine_panels = FALSE
)

Arguments

object

A `RiskMap.spatial.cv` object containing validation results.

mode

Character string specifying the mode of plotting: `"average"` (average AnPIT across test sets), `"single"` (AnPIT for a specific test set), or `"all"` (all test sets).

test_set

Integer specifying the test set to plot when `mode = "single"`.

model_name

Character string specifying the name of the model to plot. If `NULL`, all models are plotted.

combine_panels

Logical specifying whether to combine all models into a single plot when `mode = "average"`. Defaults to `FALSE`.

Value

A ggplot2 plot or a grid of plots.