Predictive Target Over a Regular Spatial Grid

Computes predictions over a regular spatial grid using outputs from the pred_over_grid function. This function allows for incorporating covariates, offsets, and optional unstructured random effects into the predictive target.

Usage

pred_target_grid(
  object,
  include_covariates = TRUE,
  include_nugget = FALSE,
  include_cov_offset = FALSE,
  include_re = FALSE,
  f_target = NULL,
  pd_summary = NULL
)

Arguments

object

Output from `pred_over_grid`, a RiskMap.pred.re object.

include_covariates

Logical. Include covariates in the predictive target.

include_nugget

Logical. Include the nugget effect in the predictive target.

include_cov_offset

Logical. Include the covariate offset in the predictive target.

include_re

Logical. Include unstructured random effects in the predictive target.

f_target

Optional. List of functions to apply on the linear predictor samples.

pd_summary

Optional. List of summary functions to apply on the predicted values.

Value

An object of class 'RiskMap_pred_target_grid' containing predicted values and summaries over the regular spatial grid.

See also

Author

Emanuele Giorgi e.giorgi@lancaster.ac.uk

Claudio Fronterre c.fronterr@lancaster.ac.uk