Compute scoring rules using spatial cross-validation

This function calculates the predictive accuracy of a spatial model fitted to a `RiskMap` object using cross-validation. It allows model scoring based on specified metrics, with options for two cross-validation methods: spatial clustering and regularized subsampling. Users can choose between continuous ranked probability score (CRPS) and scaled CRPS (SCRPS) as scoring metrics to evaluate predictive quality. For each data fold, the function can either refit the model or use fixed parameters, enabling flexible model validation and evaluation. Additionally, it can generate plots of test sets across folds, providing visual insights into the spatial cross-validation structure.

Usage

score_models(
  object,
  keep_par_fixed = TRUE,
  fold,
  n_size = NULL,
  control_sim = set_control_sim(),
  method,
  min_dist = NULL,
  plot_fold = TRUE,
  messages = TRUE,
  which_metric = "crps",
  ...
)

Arguments

object

A `RiskMap` object, output of the `glgpm` function.

keep_par_fixed

Logical; if `TRUE`, parameters are kept fixed, otherwise the model is re-estimated for each fold.

fold

Integer; number of folds for cross-validation.

n_size

Optional; the size of the test set, required if `method = "regularized"`.

control_sim

Control settings for simulation as output from `set_control_sim`.

method

Character; either `"cluster"` or `"regularized"` for the cross-validation method. The `"cluster"` method uses the spatial clustering as implented by spatial_clustering_cv functoin from the `spatialEco` package; the `"regularized"` method select a subsample of the data-set by imposing a minimum distance - as set in the argument `min_dist` - for a randomly selected subset of locations.

min_dist

Optional; minimum distance for regularized subsampling, required if `method = "regularized"`.

plot_fold

Logical; if `TRUE`, plots each fold's test set.

messages

Logical; if `TRUE`, displays progress messages.

which_metric

Character; either `"crps"` or `"scrps"` to specify scoring rule.

...

Additional arguments passed to clustering or subsampling functions.

Value

A list of class `RiskMap.spatial.cv`, containing: - `crps` or `scrps`: Computed scores for each fold. - `refit`: A list of re-fitted models for each fold.

See also

Author

Emanuele Giorgi e.giorgi@lancaster.ac.uk