Empirical variogram

Computes the empirical variogram using ``bins'' of distance provided by the user.

Usage

s_variogram(
  data,
  variable,
  bins = NULL,
  n_permutation = 0,
  convert_to_utm = TRUE,
  scale_to_km = FALSE
)

Arguments

data

an object of class sf containing the variable for which the variogram is to be computed and the coordinates

variable

a character indicating the name of variable for which the variogram is to be computed.

bins

a vector indicating the `bins` to be used to define the classes of distance used in the computation of the variogram. By default bins=NULL and bins are then computed as seq(0, d_max/2, length=15) where d_max is the maximum distance observed in the data.

n_permutation

a non-negative integer indicating the number of permutation used to compute the 95 level envelope under the assumption of spatial independence. By default n_permutation=0, and no envelope is generated.

Value

an object of class 'variogram' which is a list containing the following components

dist_class the classes of distances based on the provided values for bins

v the variogram values

n the number of pairs of observations within each bin

mid_points the middle points of the classes of distance.

lower_bound the lower values of the 95

upper_bound the upper values of the 95

Author

Emanuele Giorgi e.giorgi@lancaster.ac.uk

Claudio Fronterre c.fronterr@lancaster.ac.uk